# Conflicts:
#	src/main/java/com/superstore/services/CategoryService.java
#	src/main/java/com/superstore/services/impl/CategoryServiceImpl.java

+++ src/main/java/com/superstore/services/impl/UserServiceImpl.java
@@ -56,6 +56,11 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteById(Long id) {
+        if (dao.findById(id).isEmpty()) {
+            logger.error("User with ID {} not found", id);
+            throw new UserNotFoundException("User with ID " + id + " not found");
+        }
+
         dao.deleteById(id);
     }
 }
